Change float to double

+++ src/main/java/com/kputimcevs/insurance/common/utils/MathUtil.java
@@ -1,7 +1,7 @@
 package com.kputimcevs.insurance.common.utils;
 
 public class MathUtil {
-    public static float roundTo2DecimalPositions(float f) {
-        return (float) (Math.round(f * 100.0) / 100.0);
+    public static double roundTo2DecimalPositions(double f) {
+        return Math.round(f * 100.0) / 100.0;
     }
 }
