+++ src/test/java/com/kputimcevs/insurance/premium/PremiumCalculatorTest.java
@@ -24,36 +24,36 @@ class PremiumCalculatorTest {
 
     @Test
     public void premiumCalculatorResolvesCorrectValue() {
-        float result = premiumCalculator.calculate(policy);
-        float expected = (float) 2.28;
+        double result = premiumCalculator.calculate(policy);
+        double expected = 2.28;
         assertEquals(expected, result);
     }
 
     @Test
     public void premiumCalculatorResolvesCorrectValueWhenFireRiskExceedsThreshold() {
-        policy.policyObjects.get(0).subObjects.get(0).sumInsured = (float) 20.5;
+        policy.policyObjects.get(0).subObjects.get(0).sumInsured = 20.5;
 
-        float result = premiumCalculator.calculate(policy);
-        float expected = (float) 3.28;
+        double result = premiumCalculator.calculate(policy);
+        double expected = 3.28;
         assertEquals(expected, result);
     }
 
     @Test
     public void premiumCalculatorResolvesCorrectValueWhenTheftRiskExceedsThreshold() {
-        policy.policyObjects.get(0).subObjects.get(1).sumInsured = (float) 9.48;
+        policy.policyObjects.get(0).subObjects.get(1).sumInsured = 9.48;
 
-        float result = premiumCalculator.calculate(policy);
-        float expected = (float) 2.15;
+        double result = premiumCalculator.calculate(policy);
+        double expected = 2.15;
         assertEquals(expected, result);
     }
 
     @Test
     public void premiumCalculatorResolvesCorrectValueWhenBothRisksExceedsThreshold() {
-        policy.policyObjects.get(0).subObjects.get(0).sumInsured = (float) 20.5;
-        policy.policyObjects.get(0).subObjects.get(1).sumInsured = (float) 9.48;
+        policy.policyObjects.get(0).subObjects.get(0).sumInsured = 20.5;
+        policy.policyObjects.get(0).subObjects.get(1).sumInsured = 9.48;
 
-        float result = premiumCalculator.calculate(policy);
-        float expected = (float) 3.15;
+        double result = premiumCalculator.calculate(policy);
+        double expected = 3.15;
         assertEquals(expected, result);
     }
 }
