+++ src/test/java/com/kputimcevs/insurance/premium/services/PremiumServiceTest.java
@@ -10,7 +10,6 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import static com.kputimcevs.insurance.TestUtils.assertReflEquals;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -23,7 +22,7 @@ class PremiumServiceTest {
 
     @Test
     public void getPremiumForPolicyReturnsPremiumResponse() {
-        float premium = (float) 10.56;
+        double premium = 10.56;
         Policy policy = mock(Policy.class);
         when(premiumCalculator.calculate(policy)).thenReturn(premium);
